Check the md5 hash of a feed before deciding to parse it
It turns out that many Atom/RSS feeds these days don't implement
304 ("Not modified") responses.  In fact, more feeds seemed to get this
right 10 years ago.  So we end up wasting CPU parsing the same feed
contents over and over again.  For some large feeds, this can take
seconds.

To address this, we work out the md5 hash of a feed when we retrieve it,
and if the hash is the same as the last time it was retrieved, we
convert the response internally into a 304.